Added bookingRepo

+++ Data/Repository/BookingRepository.cs
@@ -2,9 +2,11 @@
 using FieldBooking.Data.Models;
 using FieldBooking.Domain.Models;
 using FieldBooking.Domain.Repository;
+using Microsoft.EntityFrameworkCore;
 
 namespace FieldBooking.Data.Repository
 {
+    
     public class BookingRepository : IBookingRepository
     {
 
@@ -17,27 +19,60 @@ namespace FieldBooking.Data.Repository
             _mapper = mapper;
         }
 
-        public void AddBooking(BookingDto bookingDto)
+
+        public async Task<BookingDto> CreateAsync(BookingDto bookingDto)
         {
             var booking = _mapper.Map<Booking>(bookingDto);
-            _context.Bookings.Add(booking);
-            _context.SaveChanges();
+            await _context.Bookings.AddAsync(booking);
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e.Message);
+            }
+
+            return _mapper.Map<BookingDto>(booking);
         }
 
-        public BookingDto GetBooking(int id)
+        public async Task<BookingDto> GetAsync(int id)
         {
-            return _mapper.Map<BookingDto>(_context.Bookings.FirstOrDefault(x => x.Id == id));
+            return _mapper.Map<BookingDto>(_context.Bookings.Include(x=>x.User).FirstOrDefaultAsync(x => x.Id == id));
         }
 
-        public List<BookingDto> GetAllBookings()
+        public async Task<List<BookingDto>> GetAllAsync()
+
         {
-            return _mapper.Map<List<BookingDto>>(_context.Addresses.ToList());
+            List<BookingDto> allBookings;
+            try
+            {
+                allBookings = _mapper.Map<List<BookingDto>>(_context.Bookings.ToListAsync());
+                return _mapper.Map<List<BookingDto>>(_context.Bookings.ToListAsync());
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+                throw;
+            }
         }
 
-        public void RemoveBooking(int id)
+        public async Task<BookingDto> RemoveAsync(int id)
         {
             var booking = _context.Bookings.FirstOrDefault(x => x.Id == id);
+            if (booking is null) throw new ArgumentException($"Can't find user with specified id: {id}");
             _context.Bookings.Remove(booking);
+            await _context.SaveChangesAsync();
+            return _mapper.Map<BookingDto>(booking);
+        }
+
+        public async Task<BookingDto> UpdateAsync(BookingDto bookingDto)
+        {
+            var booking = await _context.Bookings.Include(x => x.User).FirstOrDefaultAsync();
+            if (booking is null) throw new ArgumentException($"Can't find booking in database");
+            _mapper.Map(bookingDto, booking);
+            await _context.SaveChangesAsync();
+            return _mapper.Map<BookingDto>(booking);
         }
     }
 }
